+++ pyuavcan/transport/can/media/_frame.py
@@ -43,7 +43,7 @@ class DataFrame:
     @staticmethod
     def get_required_padding(data_length: int) -> int:
         """
-        Computes padding for nearest valid CAN FD frame size.
+        Computes padding to nearest valid CAN FD frame size.
 
         >>> DataFrame.get_required_padding(6)
         0
@@ -59,7 +59,7 @@ class DataFrame:
             FrameFormat.EXTENDED: "0x%08x",
             FrameFormat.BASE: "0x%03x",
         }[self.format] % self.identifier
-        return pyuavcan.util.repr_attributes(self, format=self.format.name, identifier=ide, data=self.data.hex())
+        return pyuavcan.util.repr_attributes(self, id=ide, data=self.data.hex())
 
 
 @dataclasses.dataclass(frozen=True)