+++ pyuavcan/transport/udp/__init__.py
@@ -291,7 +291,7 @@ Create two transport instances -- one with a node-ID, one anonymous:
 >>> tr_0 = pyuavcan.transport.udp.UDPTransport('127.9.1.42')
 >>> tr_0.local_node_id                                             # Derived from the IP address: (1 << 8) + 42 = 298.
 298
->>> tr_1 = pyuavcan.transport.udp.UDPTransport('127.9.15.254', anonymous=True)  # Anonymous is only for listening.
+>>> tr_1 = pyuavcan.transport.udp.UDPTransport('127.9.15.254', local_node_id=None)  # Anonymous is only for listening.
 >>> tr_1.local_node_id is None
 True
 
@@ -308,10 +308,10 @@ Send a transfer from one instance to the other:
 
 >>> await_ = tr_1.loop.run_until_complete
 >>> await_(pub.send(pyuavcan.transport.Transfer(pyuavcan.transport.Timestamp.now(),
-...                                                   pyuavcan.transport.Priority.LOW,
-...                                                   1111,
-...                                                   fragmented_payload=[]),
-...                       tr_1.loop.time() + 1.0))
+...                                             pyuavcan.transport.Priority.LOW,
+...                                             1111,
+...                                             fragmented_payload=[]),
+...                 tr_1.loop.time() + 1.0))
 True
 >>> await_(sub.receive(tr_1.loop.time() + 1.0))
 TransferFrom(..., transfer_id=1111, ...)
@@ -339,7 +339,6 @@ Inheritance diagram
                          pyuavcan.transport.udp._session._input
                          pyuavcan.transport.udp._session._output
                          pyuavcan.transport.udp._socket_reader
-                         pyuavcan.transport.udp._ip._packet
                          pyuavcan.transport.udp._tracer
    :parts: 1
 """
@@ -360,10 +359,6 @@ from ._session import UDPFeedback as UDPFeedback
 
 from ._frame import UDPFrame as UDPFrame
 
-from ._ip import MACHeader as MACHeader
-from ._ip import IPHeader as IPHeader
-from ._ip import UDPHeader as UDPHeader
-from ._ip import RawPacket as RawPacket
 from ._ip import IP_ADDRESS_NODE_ID_MASK as IP_ADDRESS_NODE_ID_MASK
 from ._ip import SUBJECT_PORT as SUBJECT_PORT
 from ._ip import node_id_to_unicast_ip as node_id_to_unicast_ip
@@ -372,7 +367,12 @@ from ._ip import message_data_specifier_to_multicast_group as message_data_speci
 from ._ip import multicast_group_to_message_data_specifier as multicast_group_to_message_data_specifier
 from ._ip import service_data_specifier_to_udp_port as service_data_specifier_to_udp_port
 from ._ip import udp_port_to_service_data_specifier as udp_port_to_service_data_specifier
+from ._ip import LinkLayerPacket as LinkLayerPacket
 
+from ._tracer import IPPacket as IPPacket
+from ._tracer import IPv4Packet as IPv4Packet
+from ._tracer import IPv6Packet as IPv6Packet
+from ._tracer import UDPIPPacket as UDPIPPacket
 from ._tracer import UDPCapture as UDPCapture
 from ._tracer import UDPTracer as UDPTracer
 from ._tracer import UDPErrorTrace as UDPErrorTrace